fix(build): extract sanitizeReturnTo to standalone module to keep server-only deps out of client bundle
